refactor: move readmodel notification classes to separate folders
