+++ backend/src/ReadModelNotifications.Tests/DefaultCommandNotificationSettingsReader/DefaultCommandNotificationSettingsReader_SagaTests.cs
@@ -5,6 +5,7 @@ using Xunit;
 using Microsoft.Extensions.Configuration;
 using FluentAssertions;
 using Common.Application.Events;
+using ReadModelNotifications.Settings;
 
 namespace ReadModelNotifications.Tests.DefaultCommandNotificationSettingsReader
 {
