+++ backend/src/ReadModelNotifications.Tests/DefaultCommandNotificationSettingsReader/DefaultCommandNotificationSettingsReader_Tests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using ReadModelNotifications.Settings;
 using System;
 using System.Collections.Generic;
 using Xunit;
