+++ backend/src/ReadModelNotifications/ExtraDataDictionaryExtensions.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using ReadModelNotifications.Settings;
 
 namespace ReadModelNotifications
 {
