+++ backend/src/ReadModelNotifications/SagaNotifications/TracedSagaNotifications.cs
@@ -1,10 +1,10 @@
-﻿using Common.Application.Commands;
+﻿using Common.Application;
+using Common.Application.Commands;
 using Common.Application.Events;
-using Common.Application.SagaNotifications;
 using Core.Common.Domain;
 using System.Diagnostics;
 
-namespace Common.Application.ReadModelNotifications
+namespace ReadModelNotifications.SagaNotifications
 {
     internal static class TracedSagaNotificationsExtensions
     {
@@ -16,34 +16,6 @@ namespace Common.Application.ReadModelNotifications
                  ((Activity?)state)?.Dispose();
                  return t;
              }, activity, TaskContinuationOptions.ExecuteSynchronously).Unwrap(); //exec synchronously to avoid calling continuation on new thread
-        }
-    }
-
-    internal class TracedImmediateNotifications : IImmediateNotifications
-    {
-        private readonly IImmediateNotifications _immediateNotifications;
-
-        public TracedImmediateNotifications(IImmediateNotifications immediateNotifications)
-        {
-            _immediateNotifications = immediateNotifications;
-        }
-
-        public Task NotifyCompleted(CorrelationId correlationId, Dictionary<string, object>? extraData = null)
-        {
-            var activity = Tracing.StartTracing(nameof(NotifyCompleted), correlationId);
-            return _immediateNotifications.NotifyCompleted(correlationId, extraData).AddActivityDisposingContinuation(activity);
-        }
-
-        public Task NotifyFailed(CorrelationId correlationId, Dictionary<string, object>? extraData = null)
-        {
-            var activity = Tracing.StartTracing(nameof(NotifyFailed), correlationId);
-            return _immediateNotifications.NotifyFailed(correlationId, extraData).AddActivityDisposingContinuation(activity);
-        }
-
-        public Task RegisterNew(CorrelationId correlationId, CommandId commandId)
-        {
-            var activity = Tracing.StartTracing(nameof(RegisterNew), correlationId);
-            return _immediateNotifications.RegisterNew(correlationId, commandId).AddActivityDisposingContinuation(activity);
         }
     }
